requirements.txt
robotframework==3.0.4
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.4.0
selenium==3.141.0
urllib3==1.25.9
